Config file structure refactored
