+++ internal/configuration/Configuration.go
@@ -1,26 +1,18 @@
 package configuration
 
 import (
-	"context"
 	"encoding/json"
-	"flag"
-	"fmt"
-	"os"
-	"path/filepath"
-	"strconv"
 	"time"
 
 	"github.com/go-co-op/gocron"
-	v1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
-	"k8s.io/client-go/rest"
-	"k8s.io/client-go/tools/clientcmd"
+
+	configTypes "github.com/tothbence9922/kawe/internal/configuration/types"
 )
 
 type Configuration struct {
-	ServerConfigs   []ServerConfiguration
-	EndpointConfigs EndpointConfiguration
+	ServerConfigs   []configTypes.ServerConfiguration
+	EndpointConfigs configTypes.EndpointConfiguration
 	scheduler       *gocron.Scheduler
 }
 
@@ -39,167 +31,8 @@ func GetInstance() *Configuration {
 	return configInstance
 }
 
-func getClientSet() *kubernetes.Clientset {
-	var clientSet *kubernetes.Clientset
-
-	// Authentication - from outside of the cluster
-	if pwd, err := os.Getwd(); err == nil && pwd != "" {
-		kubeconfig := flag.String("kubeconfig", filepath.Join(pwd, "kubeConfig.yaml"), "(optional) absolute path to the kubeconfig file")
-		// use the current context in kubeconfig
-		config, err := clientcmd.BuildConfigFromFlags("", *kubeconfig)
-
-		if err != nil {
-			panic(err.Error())
-		}
-
-		// create the clientSet
-		clientSet, err = kubernetes.NewForConfig(config)
-
-		if err != nil {
-			panic(err.Error())
-		}
-	} else {
-		// Authentication - from inside of the cluster
-		// creates the in-cluster config
-		config, err := rest.InClusterConfig()
-
-		if err != nil {
-			panic(err.Error())
-		}
-
-		// creates the clientSet
-		clientSet, err = kubernetes.NewForConfig(config)
-
-		if err != nil {
-			panic(err.Error())
-		}
-	}
-	return clientSet
-}
-
-func getAnnotationsForEndpointByName(name string, services *v1.ServiceList) map[string]string {
-	for _, service := range services.Items {
-		if service.ObjectMeta.Name == name {
-			return service.Annotations
-		}
-	}
-	return make(map[string]string)
-}
-
-func getNameSpaceConfigs(clientSet *kubernetes.Clientset) []NamespaceConfiguration {
-	namespaces, err := clientSet.CoreV1().Namespaces().List(context.TODO(), metav1.ListOptions{})
-
-	if err != nil {
-		panic(err.Error())
-	}
-
-	namespaceConfigs := []NamespaceConfiguration{}
-
-	// Iterating through all namespaces, adding services and pods behind services to the configuration
-	for _, namespace := range namespaces.Items {
-
-		// Getting all pods - including the ones behind services and separate ones aswell.
-		allPods, err := clientSet.CoreV1().Pods(namespace.ObjectMeta.Name).List(context.TODO(), metav1.ListOptions{})
-
-		if err != nil {
-			panic(err.Error())
-		}
-
-		podConfigs := []PodConfiguration{}
-
-		for _, pod := range allPods.Items {
-			podIp := pod.Status.PodIP
-			podName := pod.ObjectMeta.Name
-			podLabels := pod.ObjectMeta.Labels
-			podAnnotations := pod.GetAnnotations()
-			// TODO use target port from service.Spec.Ports
-			podConfigs = append(podConfigs, PodConfiguration{Address: podIp, Name: podName, Labels: podLabels, Annotations: podAnnotations, Enabled: true, Periodicity: 5, Timeout: 5000})
-		}
-
-		// Using endpoints as "services" to see IP addresses
-		endpoints, err := clientSet.CoreV1().Endpoints(namespace.ObjectMeta.Name).List(context.TODO(), metav1.ListOptions{})
-		services, err := clientSet.CoreV1().Services(namespace.ObjectMeta.Name).List(context.TODO(), metav1.ListOptions{})
-
-		if err != nil {
-			panic(err.Error())
-		}
-
-		serviceConfigurations := []ServiceConfiguration{}
-
-		for _, endpoint := range endpoints.Items {
-
-			serviceName := endpoint.ObjectMeta.Name
-			serviceAnnotations := getAnnotationsForEndpointByName(endpoint.ObjectMeta.Name, services)
-			currentPodConfigs := []PodConfiguration{}
-
-			for _, subset := range endpoint.Subsets {
-				for _, address := range subset.Addresses {
-					for _, podConfig := range podConfigs {
-						if address.IP == podConfig.Address {
-							currentPodConfigs = append(currentPodConfigs, podConfig)
-						}
-					}
-				}
-			}
-
-			serviceConfigurations = append(serviceConfigurations, ServiceConfiguration{Name: serviceName, Annotations: serviceAnnotations, Pods: currentPodConfigs})
-
-		}
-
-		separatePodConfigs := []PodConfiguration{}
-
-		for _, podConfig := range podConfigs {
-			found := false
-
-			for _, serviceConfig := range serviceConfigurations {
-				for _, svcPodConfig := range serviceConfig.Pods {
-					if svcPodConfig.Address == podConfig.Address {
-						found = true
-					}
-				}
-			}
-
-			if !found {
-				separatePodConfigs = append(separatePodConfigs, podConfig)
-			}
-		}
-
-		namespaceConfigs = append(namespaceConfigs, NamespaceConfiguration{Name: namespace.ObjectMeta.Name, Services: serviceConfigurations, Pods: separatePodConfigs})
-	}
-
-	return namespaceConfigs
-}
-
-func getServerConfigurations() []ServerConfiguration {
-	var serverConfigs []ServerConfiguration
-
-	httpPortString := os.Getenv("KAWE_HTTP_PORT")
-	if httpPortString != "" {
-		httpPort, err := strconv.Atoi(httpPortString)
-		if err == nil {
-			serverConfigs = append(serverConfigs, ServerConfiguration{Type: "HTTP", Port: httpPort})
-		}
-	} else {
-		fmt.Println("KAWE_HTTP_PORT not found, using default port")
-		serverConfigs = append(serverConfigs, ServerConfiguration{Type: "HTTP", Port: 80})
-	}
-
-	prometheusPortString := os.Getenv("KAWE_PROMETHEUS_PORT")
-	if prometheusPortString != "" {
-		prometheusPort, err := strconv.Atoi(prometheusPortString)
-		if err == nil {
-			serverConfigs = append(serverConfigs, ServerConfiguration{Type: "PROMETHEUS", Port: prometheusPort})
-		}
-	} else {
-		fmt.Println("KAWE_PROMETHEUS_PORT not found, using default port")
-		serverConfigs = append(serverConfigs, ServerConfiguration{Type: "PROMETHEUS", Port: 80})
-	}
-
-	return serverConfigs
-}
-
 func (cfg *Configuration) getTargets(clientSet *kubernetes.Clientset) {
-	cfg.EndpointConfigs = EndpointConfiguration{Namespaces: getNameSpaceConfigs(clientSet)}
+	cfg.EndpointConfigs = configTypes.EndpointConfiguration{Namespaces: getNameSpaceConfigs(clientSet)}
 }
 
 func (cfg *Configuration) GetConfiguration() {
@@ -222,83 +55,4 @@ func (c Configuration) String() string {
 	}
 
 	return ret
-}
-
-type ServerConfiguration struct {
-	Type string
-	Port int
-}
-
-func (sc ServerConfiguration) String() string {
-
-	portString := fmt.Sprintf(":%d", sc.Port)
-
-	return sc.Type + portString
-}
-
-type EndpointConfiguration struct {
-	Namespaces []NamespaceConfiguration
-}
-
-type NamespaceConfiguration struct {
-	Name     string
-	Services []ServiceConfiguration
-	Pods     []PodConfiguration
-}
-
-type ServiceConfiguration struct {
-	Name            string
-	Annotations     map[string]string
-	Pods            []PodConfiguration
-	ProcessorConfig ProcessorConfiguration
-}
-
-type PodConfiguration struct {
-	Name        string
-	Labels      map[string]string
-	Annotations map[string]string
-	Address     string
-	Enabled     bool
-	Periodicity int
-	Timeout     int
-}
-
-func (p PodConfiguration) String() string {
-	enabledString := "DISABLED"
-	if p.Enabled {
-		enabledString = "ENABLED"
-	}
-	ret := p.Name + " " + p.Address + " " + enabledString + " " + fmt.Sprint(p.Periodicity) + " " + fmt.Sprint(p.Timeout)
-	return ret
-}
-
-func (sc ServiceConfiguration) String() string {
-
-	ret := ""
-	for _, curPingConfig := range sc.Pods {
-		ret += curPingConfig.String()
-	}
-	return ret
-}
-
-type PingConfiguration struct {
-	Periodicity int
-	Method      interface{}
-	Target      string
-	Timeout     int
-}
-
-func (pc PingConfiguration) String() string {
-
-	return fmt.Sprintf("periodicity\t\ttarget\n%d\t\t%s\n", pc.Periodicity, pc.Target)
-}
-
-type ProcessorConfiguration struct {
-	Type   string
-	Params interface{}
-}
-
-func (pc ProcessorConfiguration) String() string {
-
-	return fmt.Sprintf("\tType\n%s\n", pc.Type)
 }
