+++ internal/configuration/types/EndpointConfiguration.go
@@ -1,0 +1,5 @@
+package types
+
+type EndpointConfiguration struct {
+	Namespaces []NamespaceConfiguration
+}
