+++ internal/configuration/types/NamespaceConfiguration.go
@@ -1,0 +1,7 @@
+package types
+
+type NamespaceConfiguration struct {
+	Name     string
+	Services []ServiceConfiguration
+	Pods     []PodConfiguration
+}
