+++ internal/configuration/types/PingConfiguration.go
@@ -1,0 +1,15 @@
+package types
+
+import "fmt"
+
+type PingConfiguration struct {
+	Periodicity int
+	Method      interface{}
+	Target      string
+	Timeout     int
+}
+
+func (pc PingConfiguration) String() string {
+
+	return fmt.Sprintf("periodicity\t\ttarget\n%d\t\t%s\n", pc.Periodicity, pc.Target)
+}
