+++ internal/configuration/types/ProcessorConfiguration.go
@@ -1,0 +1,13 @@
+package types
+
+import "fmt"
+
+type ProcessorConfiguration struct {
+	Type   string
+	Params interface{}
+}
+
+func (pc ProcessorConfiguration) String() string {
+
+	return fmt.Sprintf("\tType\n%s\n", pc.Type)
+}
