+++ internal/configuration/types/ServerConfiguration.go
@@ -1,0 +1,15 @@
+package types
+
+import "fmt"
+
+type ServerConfiguration struct {
+	Type string
+	Port int
+}
+
+func (sc ServerConfiguration) String() string {
+
+	portString := fmt.Sprintf(":%d", sc.Port)
+
+	return sc.Type + portString
+}
