+++ internal/configuration/types/ServiceConfiguration.go
@@ -1,0 +1,17 @@
+package types
+
+type ServiceConfiguration struct {
+	Name            string
+	Annotations     map[string]string
+	Pods            []PodConfiguration
+	ProcessorConfig ProcessorConfiguration
+}
+
+func (sc ServiceConfiguration) String() string {
+
+	ret := ""
+	for _, curPingConfig := range sc.Pods {
+		ret += curPingConfig.String()
+	}
+	return ret
+}
